refactor poi editor complexity hotspots
