+++ packages/plugin-poi/src/plugin.translations.de.ts
@@ -1,0 +1,183 @@
+import { createPoiLocaleTranslations } from './plugin.translations.shared.js';
+
+export const poiTranslationsDe = createPoiLocaleTranslations({
+  navigation: { title: 'POI' },
+  list: { title: 'POI', description: 'Points of Interest aus dem Mainserver bearbeiten.' },
+  detail: {
+    createTitle: 'POI anlegen',
+    createDescription: 'Erstellen Sie einen neuen Point of Interest.',
+    editTitle: 'POI bearbeiten',
+    editDescription: 'Aktualisieren oder löschen Sie den Point of Interest.',
+  },
+  editor: {
+    createTitle: 'POI anlegen',
+    createDescription: 'Erstellen Sie einen neuen Point of Interest.',
+    editTitle: 'POI bearbeiten',
+    editDescription: 'Aktualisieren oder löschen Sie den Point of Interest.',
+  },
+  fields: {
+    name: 'Name',
+    description: 'Beschreibung',
+    mobileDescription: 'Mobile Beschreibung',
+    teaserImage: 'Teaserbild',
+    active: 'Aktiv',
+    categoryName: 'Kategorie',
+    street: 'Straße',
+    zip: 'PLZ',
+    city: 'Ort',
+    latitude: 'Breitengrad',
+    longitude: 'Längengrad',
+    locationName: 'Ortsbezeichnung',
+    addressSearch: 'Adresse suchen',
+    searchResults: 'Suchergebnisse',
+    contact: 'Kontakt',
+    firstName: 'Vorname',
+    lastName: 'Nachname',
+    phone: 'Telefon',
+    email: 'E-Mail',
+    operatorName: 'Betreiber',
+    url: 'Web-URL',
+    urlDescription: 'Link-Beschreibung',
+    mediaPlaceholder: 'Medium auswählen',
+    assetId: 'Asset-ID',
+    attachmentAsset: 'Weiteres Medium',
+    label: 'Bezeichnung',
+    weekday: 'Wochentag',
+    timeFrom: 'Öffnet',
+    timeTo: 'Schließt',
+    open: 'Geöffnet',
+    priceName: 'Preisname',
+    amount: 'Betrag',
+    certificateName: 'Zertifikat',
+    accessibilityDescription: 'Barrierefreiheit',
+    accessibilityTypes: 'Barrierefreiheits-Typen',
+    tags: 'Tags',
+    payload: 'Payload JSON',
+    actions: 'Aktionen',
+    createdAt: 'Erstellt',
+    updatedAt: 'Aktualisiert',
+  },
+  actions: {
+    create: 'Anlegen',
+    update: 'Änderungen speichern',
+    save: 'Speichern',
+    edit: 'Bearbeiten',
+    delete: 'Löschen',
+    back: 'Zurück zur Liste',
+    deleteConfirm: 'Soll dieser POI wirklich gelöscht werden?',
+    clearMedia: 'Medium entfernen',
+    add: 'Hinzufügen',
+    remove: 'Entfernen',
+    searchAddress: 'Adresse suchen',
+    applySearchResult: 'Übernehmen',
+    reverseGeocode: 'Adresse aus Koordinaten übernehmen',
+    reverseGeocoding: 'Adresse wird ermittelt',
+    uploadMedia: 'Medium hochladen',
+    uploadingMedia: 'Medium wird hochgeladen',
+  },
+  messages: {
+    loading: 'POI werden geladen.',
+    loadError: 'POI konnten nicht geladen werden.',
+    missingContent: 'Der POI konnte nicht geladen werden.',
+    saveError: 'POI konnte nicht gespeichert werden.',
+    deleteError: 'POI konnte nicht gelöscht werden.',
+    createSuccess: 'POI wurde erstellt.',
+    updateSuccess: 'POI wurde aktualisiert.',
+    validationError: 'Bitte korrigieren Sie die markierten Felder.',
+    locationSearchError: 'Die Adresssuche ist derzeit nicht verfügbar. Sie können die Felder manuell bearbeiten.',
+    locationSearchEmpty: 'Es wurden keine Adresstreffer gefunden.',
+    locationReverseGeocodeError: 'Die Adressauflösung aus Koordinaten ist derzeit nicht verfügbar.',
+    locationReverseGeocodeEmpty: 'Zu diesen Koordinaten wurde keine Adresse gefunden.',
+    locationMapUnavailable: 'Die Kartenansicht ist für diese Instanz derzeit deaktiviert.',
+    locationMapError: 'Die Kartenansicht konnte nicht geladen werden. Die Felder bleiben manuell bearbeitbar.',
+    mediaUploadError: 'Das Medium konnte nicht hochgeladen werden.',
+    mediaUploadSuccess: 'Das Medium wurde hochgeladen und kann jetzt referenziert werden.',
+  },
+  tabs: {
+    ariaLabel: 'Detailbereiche',
+    mobileLabel: 'Bereich auswählen',
+  },
+  detailTabs: {
+    basis: { title: 'Basis', description: 'Stammdaten und Veröffentlichungsstatus.' },
+    location: { title: 'Ort', description: 'Adresse, Ortsdaten und Geo-Koordinaten.' },
+    description: { title: 'Beschreibung', description: 'Redaktionelle Beschreibungen des POI.' },
+    contact: { title: 'Kontakt', description: 'Primäre Kontaktinformationen.' },
+    openingHours: { title: 'Öffnungszeiten', description: 'Wiederholbare Öffnungsinformationen.' },
+    links: { title: 'Links', description: 'Externe Verweise und weiterführende URLs.' },
+    operator: { title: 'Betreiber', description: 'Betreiberdaten und zuständige Organisation.' },
+    prices: { title: 'Preise', description: 'Preisangaben und Kategorien.' },
+    media: { title: 'Medien & Dateien', description: 'Teaserbild und weitere Dateireferenzen.' },
+    advanced: { title: 'Erweiterte Daten', description: 'Payload, Zertifikate und Barrierefreiheit.' },
+    history: { title: 'Historie', description: 'Änderungs- und Aktivitätsverlauf.' },
+  },
+  cards: {
+    basis: {
+      identity: { title: 'Basisdaten', description: 'Name, Kategorie und Aktivstatus.' },
+      meta: { title: 'Metadaten', description: 'Zeitliche Einordnung des Eintrags.' },
+    },
+    content: {
+      descriptions: { title: 'Beschreibungen', description: 'Redaktionelle Beschreibungen des POI.' },
+      contact: { title: 'Kontakt', description: 'Kontaktinformationen für den POI.' },
+      location: { title: 'Lage und Adresse', description: 'Adressdaten des POI.' },
+      openingHours: { title: 'Öffnungszeiten', description: 'Aktuelle Öffnungsinformationen.' },
+      links: { title: 'Weblinks', description: 'Externe Verweise zum POI.' },
+      payload: { title: 'Zusatzdaten', description: 'Zusätzliche Mainserver-Daten als JSON.' },
+    },
+    location: {
+      search: { title: 'Adresssuche', description: 'Adresse suchen und in den POI übernehmen.' },
+      address: { title: 'Lage und Adresse', description: 'Strukturierte Adressdaten des POI.' },
+      coordinates: { title: 'Koordinaten', description: 'Geo-Koordinaten des POI.' },
+      map: { title: 'Karte', description: 'Marker setzen oder verschieben, um Koordinaten zu pflegen.' },
+    },
+    description: {
+      text: { title: 'Beschreibungen', description: 'Redaktionelle Texte für den POI.' },
+    },
+    contact: {
+      primary: { title: 'Kontakt', description: 'Kontaktinformationen für den POI.' },
+    },
+    openingHours: {
+      entries: { title: 'Öffnungszeiten', description: 'Mehrere Öffnungszeiten pflegen.' },
+    },
+    links: {
+      entries: { title: 'Weblinks', description: 'Externe Verweise zum POI.' },
+    },
+    operator: {
+      details: { title: 'Betreiber', description: 'Zuständige Organisation und Erreichbarkeit.' },
+    },
+    prices: {
+      entries: { title: 'Preise', description: 'Mehrere Preisangaben pflegen.' },
+    },
+    media: {
+      references: { title: 'Medien', description: 'Teaserbild und Dateireferenzen.' },
+    },
+    advanced: {
+      payload: { title: 'Zusatzdaten', description: 'Payload, Zertifikate und Barrierefreiheit.' },
+    },
+    settings: {
+      media: { title: 'Medien', description: 'Teaserbild für die Übersicht.' },
+    },
+  },
+  history: {
+    empty: {
+      title: 'Noch keine Historie verfügbar.',
+      description: 'Historienereignisse für POI werden in einem späteren Schritt angebunden.',
+    },
+  },
+  empty: { title: 'Noch keine POI vorhanden', description: 'Legen Sie den ersten POI an.' },
+  pagination: {
+    ariaLabel: 'POI-Pagination',
+    previous: 'Zurück',
+    next: 'Weiter',
+    pageLabel: 'Seite {{page}}',
+  },
+  values: {
+    notAvailable: 'Nicht verfügbar',
+    active: 'Ja',
+  },
+  validation: {
+    name: 'Der Name ist erforderlich.',
+    webUrls: 'URLs müssen mit https:// beginnen.',
+    categoryName: 'Die Kategorie darf maximal 128 Zeichen haben.',
+    payload: 'Payload muss gültiges JSON sein.',
+  },
+});
