+++ packages/plugin-poi/src/plugin.translations.en.ts
@@ -1,0 +1,183 @@
+import { createPoiLocaleTranslations } from './plugin.translations.shared.js';
+
+export const poiTranslationsEn = createPoiLocaleTranslations({
+  navigation: { title: 'POI' },
+  list: { title: 'POI', description: 'Edit Mainserver points of interest.' },
+  detail: {
+    createTitle: 'Create POI',
+    createDescription: 'Create a new point of interest.',
+    editTitle: 'Edit POI',
+    editDescription: 'Update or delete the point of interest.',
+  },
+  editor: {
+    createTitle: 'Create POI',
+    createDescription: 'Create a new point of interest.',
+    editTitle: 'Edit POI',
+    editDescription: 'Update or delete the point of interest.',
+  },
+  fields: {
+    name: 'Name',
+    description: 'Description',
+    mobileDescription: 'Mobile description',
+    teaserImage: 'Teaser image',
+    active: 'Active',
+    categoryName: 'Category',
+    street: 'Street',
+    zip: 'ZIP',
+    city: 'City',
+    latitude: 'Latitude',
+    longitude: 'Longitude',
+    locationName: 'Location name',
+    addressSearch: 'Search address',
+    searchResults: 'Search results',
+    contact: 'Contact',
+    firstName: 'First name',
+    lastName: 'Last name',
+    phone: 'Phone',
+    email: 'Email',
+    operatorName: 'Operator',
+    url: 'Web URL',
+    urlDescription: 'Link description',
+    mediaPlaceholder: 'Select media',
+    assetId: 'Asset ID',
+    attachmentAsset: 'Additional media',
+    label: 'Label',
+    weekday: 'Weekday',
+    timeFrom: 'Opens',
+    timeTo: 'Closes',
+    open: 'Open',
+    priceName: 'Price name',
+    amount: 'Amount',
+    certificateName: 'Certificate',
+    accessibilityDescription: 'Accessibility',
+    accessibilityTypes: 'Accessibility types',
+    tags: 'Tags',
+    payload: 'Payload JSON',
+    actions: 'Actions',
+    createdAt: 'Created',
+    updatedAt: 'Updated',
+  },
+  actions: {
+    create: 'Create',
+    update: 'Save changes',
+    save: 'Save',
+    edit: 'Edit',
+    delete: 'Delete',
+    back: 'Back to list',
+    deleteConfirm: 'Delete this POI?',
+    clearMedia: 'Clear media',
+    add: 'Add',
+    remove: 'Remove',
+    searchAddress: 'Search address',
+    applySearchResult: 'Apply',
+    reverseGeocode: 'Resolve address from coordinates',
+    reverseGeocoding: 'Resolving address',
+    uploadMedia: 'Upload media',
+    uploadingMedia: 'Uploading media',
+  },
+  messages: {
+    loading: 'Loading POI.',
+    loadError: 'POI could not be loaded.',
+    missingContent: 'The POI could not be loaded.',
+    saveError: 'POI could not be saved.',
+    deleteError: 'POI could not be deleted.',
+    createSuccess: 'POI was created.',
+    updateSuccess: 'POI was updated.',
+    validationError: 'Please correct the highlighted fields.',
+    locationSearchError: 'Address search is currently unavailable. You can still edit the fields manually.',
+    locationSearchEmpty: 'No address matches were found.',
+    locationReverseGeocodeError: 'Reverse geocoding is currently unavailable.',
+    locationReverseGeocodeEmpty: 'No address was found for these coordinates.',
+    locationMapUnavailable: 'The map view is currently disabled for this instance.',
+    locationMapError: 'The map view could not be loaded. Manual editing remains available.',
+    mediaUploadError: 'The media upload failed.',
+    mediaUploadSuccess: 'The media was uploaded and can now be referenced.',
+  },
+  tabs: {
+    ariaLabel: 'Detail sections',
+    mobileLabel: 'Select section',
+  },
+  detailTabs: {
+    basis: { title: 'Basics', description: 'Master data and publication status.' },
+    location: { title: 'Location', description: 'Address, location data, and geo coordinates.' },
+    description: { title: 'Description', description: 'Editorial description fields.' },
+    contact: { title: 'Contact', description: 'Primary contact details.' },
+    openingHours: { title: 'Opening hours', description: 'Repeatable opening entries.' },
+    links: { title: 'Links', description: 'External references and related URLs.' },
+    operator: { title: 'Operator', description: 'Operating company and responsibility.' },
+    prices: { title: 'Prices', description: 'Price information and categories.' },
+    media: { title: 'Media & files', description: 'Teaser image and file references.' },
+    advanced: { title: 'Advanced data', description: 'Payload, certificates, and accessibility.' },
+    history: { title: 'History', description: 'Change and activity history.' },
+  },
+  cards: {
+    basis: {
+      identity: { title: 'Core data', description: 'Name, category, and activation state.' },
+      meta: { title: 'Metadata', description: 'Temporal information for the record.' },
+    },
+    content: {
+      descriptions: { title: 'Descriptions', description: 'Editorial descriptions for the POI.' },
+      contact: { title: 'Contact', description: 'Contact information for the POI.' },
+      location: { title: 'Location and address', description: 'Address data of the POI.' },
+      openingHours: { title: 'Opening hours', description: 'Current opening information.' },
+      links: { title: 'Web links', description: 'External references for the POI.' },
+      payload: { title: 'Additional data', description: 'Additional Mainserver data as JSON.' },
+    },
+    location: {
+      search: { title: 'Address search', description: 'Search an address and apply it to the POI.' },
+      address: { title: 'Location and address', description: 'Structured address data for the POI.' },
+      coordinates: { title: 'Coordinates', description: 'Geo coordinates of the POI.' },
+      map: { title: 'Map', description: 'Place or drag the marker to maintain coordinates.' },
+    },
+    description: {
+      text: { title: 'Descriptions', description: 'Editorial text fields for the POI.' },
+    },
+    contact: {
+      primary: { title: 'Contact', description: 'Contact information for the POI.' },
+    },
+    openingHours: {
+      entries: { title: 'Opening hours', description: 'Maintain multiple opening entries.' },
+    },
+    links: {
+      entries: { title: 'Web links', description: 'External references for the POI.' },
+    },
+    operator: {
+      details: { title: 'Operator', description: 'Responsible organization and reachability.' },
+    },
+    prices: {
+      entries: { title: 'Prices', description: 'Maintain multiple price entries.' },
+    },
+    media: {
+      references: { title: 'Media', description: 'Teaser image and file references.' },
+    },
+    advanced: {
+      payload: { title: 'Additional data', description: 'Payload, certificates, and accessibility.' },
+    },
+    settings: {
+      media: { title: 'Media', description: 'Teaser image for the overview.' },
+    },
+  },
+  history: {
+    empty: {
+      title: 'No history available yet.',
+      description: 'POI history entries will be connected in a later step.',
+    },
+  },
+  empty: { title: 'No POI yet', description: 'Create the first POI.' },
+  pagination: {
+    ariaLabel: 'POI pagination',
+    previous: 'Previous',
+    next: 'Next',
+    pageLabel: 'Page {{page}}',
+  },
+  values: {
+    notAvailable: 'Not available',
+    active: 'Yes',
+  },
+  validation: {
+    name: 'Name is required.',
+    webUrls: 'URLs must start with https://.',
+    categoryName: 'Category must be at most 128 characters.',
+    payload: 'Payload must be valid JSON.',
+  },
+});
