+++ packages/plugin-poi/src/plugin.translations.shared.ts
@@ -1,0 +1,2 @@
+export const createPoiLocaleTranslations = <TPoi extends Readonly<Record<string, unknown>>>(poi: TPoi) =>
+  ({ poi }) as const;
