+++ packages/plugin-poi/src/plugin.translations.ts
@@ -1,371 +1,7 @@
 import type { PluginTranslations } from '@sva/plugin-sdk';
 
-const createPoiLocaleTranslations = <TPoi extends Readonly<Record<string, unknown>>>(poi: TPoi) =>
-  ({ poi }) as const;
-
-const poiTranslationsDe = createPoiLocaleTranslations({
-  navigation: { title: 'POI' },
-  list: { title: 'POI', description: 'Points of Interest aus dem Mainserver bearbeiten.' },
-  detail: {
-    createTitle: 'POI anlegen',
-    createDescription: 'Erstellen Sie einen neuen Point of Interest.',
-    editTitle: 'POI bearbeiten',
-    editDescription: 'Aktualisieren oder löschen Sie den Point of Interest.',
-  },
-  editor: {
-    createTitle: 'POI anlegen',
-    createDescription: 'Erstellen Sie einen neuen Point of Interest.',
-    editTitle: 'POI bearbeiten',
-    editDescription: 'Aktualisieren oder löschen Sie den Point of Interest.',
-  },
-  fields: {
-    name: 'Name',
-    description: 'Beschreibung',
-    mobileDescription: 'Mobile Beschreibung',
-    teaserImage: 'Teaserbild',
-    active: 'Aktiv',
-    categoryName: 'Kategorie',
-    street: 'Straße',
-    zip: 'PLZ',
-    city: 'Ort',
-    latitude: 'Breitengrad',
-    longitude: 'Längengrad',
-    locationName: 'Ortsbezeichnung',
-    addressSearch: 'Adresse suchen',
-    searchResults: 'Suchergebnisse',
-    contact: 'Kontakt',
-    firstName: 'Vorname',
-    lastName: 'Nachname',
-    phone: 'Telefon',
-    email: 'E-Mail',
-    operatorName: 'Betreiber',
-    url: 'Web-URL',
-    urlDescription: 'Link-Beschreibung',
-    mediaPlaceholder: 'Medium auswählen',
-    assetId: 'Asset-ID',
-    attachmentAsset: 'Weiteres Medium',
-    label: 'Bezeichnung',
-    weekday: 'Wochentag',
-    timeFrom: 'Öffnet',
-    timeTo: 'Schließt',
-    open: 'Geöffnet',
-    priceName: 'Preisname',
-    amount: 'Betrag',
-    certificateName: 'Zertifikat',
-    accessibilityDescription: 'Barrierefreiheit',
-    accessibilityTypes: 'Barrierefreiheits-Typen',
-    tags: 'Tags',
-    payload: 'Payload JSON',
-    actions: 'Aktionen',
-    createdAt: 'Erstellt',
-    updatedAt: 'Aktualisiert',
-  },
-  actions: {
-    create: 'Anlegen',
-    update: 'Änderungen speichern',
-    save: 'Speichern',
-    edit: 'Bearbeiten',
-    delete: 'Löschen',
-    back: 'Zurück zur Liste',
-    deleteConfirm: 'Soll dieser POI wirklich gelöscht werden?',
-    clearMedia: 'Medium entfernen',
-    add: 'Hinzufügen',
-    remove: 'Entfernen',
-    searchAddress: 'Adresse suchen',
-    applySearchResult: 'Übernehmen',
-    reverseGeocode: 'Adresse aus Koordinaten übernehmen',
-    reverseGeocoding: 'Adresse wird ermittelt',
-    uploadMedia: 'Medium hochladen',
-    uploadingMedia: 'Medium wird hochgeladen',
-  },
-  messages: {
-    loading: 'POI werden geladen.',
-    loadError: 'POI konnten nicht geladen werden.',
-    missingContent: 'Der POI konnte nicht geladen werden.',
-    saveError: 'POI konnte nicht gespeichert werden.',
-    deleteError: 'POI konnte nicht gelöscht werden.',
-    createSuccess: 'POI wurde erstellt.',
-    updateSuccess: 'POI wurde aktualisiert.',
-    validationError: 'Bitte korrigieren Sie die markierten Felder.',
-    locationSearchError: 'Die Adresssuche ist derzeit nicht verfügbar. Sie können die Felder manuell bearbeiten.',
-    locationSearchEmpty: 'Es wurden keine Adresstreffer gefunden.',
-    locationReverseGeocodeError: 'Die Adressauflösung aus Koordinaten ist derzeit nicht verfügbar.',
-    locationReverseGeocodeEmpty: 'Zu diesen Koordinaten wurde keine Adresse gefunden.',
-    locationMapUnavailable: 'Die Kartenansicht ist für diese Instanz derzeit deaktiviert.',
-    locationMapError: 'Die Kartenansicht konnte nicht geladen werden. Die Felder bleiben manuell bearbeitbar.',
-    mediaUploadError: 'Das Medium konnte nicht hochgeladen werden.',
-    mediaUploadSuccess: 'Das Medium wurde hochgeladen und kann jetzt referenziert werden.',
-  },
-  tabs: {
-    ariaLabel: 'Detailbereiche',
-    mobileLabel: 'Bereich auswählen',
-  },
-  detailTabs: {
-    basis: { title: 'Basis', description: 'Stammdaten und Veröffentlichungsstatus.' },
-    location: { title: 'Ort', description: 'Adresse, Ortsdaten und Geo-Koordinaten.' },
-    description: { title: 'Beschreibung', description: 'Redaktionelle Beschreibungen des POI.' },
-    contact: { title: 'Kontakt', description: 'Primäre Kontaktinformationen.' },
-    openingHours: { title: 'Öffnungszeiten', description: 'Wiederholbare Öffnungsinformationen.' },
-    links: { title: 'Links', description: 'Externe Verweise und weiterführende URLs.' },
-    operator: { title: 'Betreiber', description: 'Betreiberdaten und zuständige Organisation.' },
-    prices: { title: 'Preise', description: 'Preisangaben und Kategorien.' },
-    media: { title: 'Medien & Dateien', description: 'Teaserbild und weitere Dateireferenzen.' },
-    advanced: { title: 'Erweiterte Daten', description: 'Payload, Zertifikate und Barrierefreiheit.' },
-    history: { title: 'Historie', description: 'Änderungs- und Aktivitätsverlauf.' },
-  },
-  cards: {
-    basis: {
-      identity: { title: 'Basisdaten', description: 'Name, Kategorie und Aktivstatus.' },
-      meta: { title: 'Metadaten', description: 'Zeitliche Einordnung des Eintrags.' },
-    },
-    content: {
-      descriptions: { title: 'Beschreibungen', description: 'Redaktionelle Beschreibungen des POI.' },
-      contact: { title: 'Kontakt', description: 'Kontaktinformationen für den POI.' },
-      location: { title: 'Lage und Adresse', description: 'Adressdaten des POI.' },
-      openingHours: { title: 'Öffnungszeiten', description: 'Aktuelle Öffnungsinformationen.' },
-      links: { title: 'Weblinks', description: 'Externe Verweise zum POI.' },
-      payload: { title: 'Zusatzdaten', description: 'Zusätzliche Mainserver-Daten als JSON.' },
-    },
-    location: {
-      search: { title: 'Adresssuche', description: 'Adresse suchen und in den POI übernehmen.' },
-      address: { title: 'Lage und Adresse', description: 'Strukturierte Adressdaten des POI.' },
-      coordinates: { title: 'Koordinaten', description: 'Geo-Koordinaten des POI.' },
-      map: { title: 'Karte', description: 'Marker setzen oder verschieben, um Koordinaten zu pflegen.' },
-    },
-    description: {
-      text: { title: 'Beschreibungen', description: 'Redaktionelle Texte für den POI.' },
-    },
-    contact: {
-      primary: { title: 'Kontakt', description: 'Kontaktinformationen für den POI.' },
-    },
-    openingHours: {
-      entries: { title: 'Öffnungszeiten', description: 'Mehrere Öffnungszeiten pflegen.' },
-    },
-    links: {
-      entries: { title: 'Weblinks', description: 'Externe Verweise zum POI.' },
-    },
-    operator: {
-      details: { title: 'Betreiber', description: 'Zuständige Organisation und Erreichbarkeit.' },
-    },
-    prices: {
-      entries: { title: 'Preise', description: 'Mehrere Preisangaben pflegen.' },
-    },
-    media: {
-      references: { title: 'Medien', description: 'Teaserbild und Dateireferenzen.' },
-    },
-    advanced: {
-      payload: { title: 'Zusatzdaten', description: 'Payload, Zertifikate und Barrierefreiheit.' },
-    },
-    settings: {
-      media: { title: 'Medien', description: 'Teaserbild für die Übersicht.' },
-    },
-  },
-  history: {
-    empty: {
-      title: 'Noch keine Historie verfügbar.',
-      description: 'Historienereignisse für POI werden in einem späteren Schritt angebunden.',
-    },
-  },
-  empty: { title: 'Noch keine POI vorhanden', description: 'Legen Sie den ersten POI an.' },
-  pagination: {
-    ariaLabel: 'POI-Pagination',
-    previous: 'Zurück',
-    next: 'Weiter',
-    pageLabel: 'Seite {{page}}',
-  },
-  values: {
-    notAvailable: 'Nicht verfügbar',
-    active: 'Ja',
-  },
-  validation: {
-    name: 'Der Name ist erforderlich.',
-    webUrls: 'URLs müssen mit https:// beginnen.',
-    categoryName: 'Die Kategorie darf maximal 128 Zeichen haben.',
-    payload: 'Payload muss gültiges JSON sein.',
-  },
-});
-
-const poiTranslationsEn = createPoiLocaleTranslations({
-  navigation: { title: 'POI' },
-  list: { title: 'POI', description: 'Edit Mainserver points of interest.' },
-  detail: {
-    createTitle: 'Create POI',
-    createDescription: 'Create a new point of interest.',
-    editTitle: 'Edit POI',
-    editDescription: 'Update or delete the point of interest.',
-  },
-  editor: {
-    createTitle: 'Create POI',
-    createDescription: 'Create a new point of interest.',
-    editTitle: 'Edit POI',
-    editDescription: 'Update or delete the point of interest.',
-  },
-  fields: {
-    name: 'Name',
-    description: 'Description',
-    mobileDescription: 'Mobile description',
-    teaserImage: 'Teaser image',
-    active: 'Active',
-    categoryName: 'Category',
-    street: 'Street',
-    zip: 'ZIP',
-    city: 'City',
-    latitude: 'Latitude',
-    longitude: 'Longitude',
-    locationName: 'Location name',
-    addressSearch: 'Search address',
-    searchResults: 'Search results',
-    contact: 'Contact',
-    firstName: 'First name',
-    lastName: 'Last name',
-    phone: 'Phone',
-    email: 'Email',
-    operatorName: 'Operator',
-    url: 'Web URL',
-    urlDescription: 'Link description',
-    mediaPlaceholder: 'Select media',
-    assetId: 'Asset ID',
-    attachmentAsset: 'Additional media',
-    label: 'Label',
-    weekday: 'Weekday',
-    timeFrom: 'Opens',
-    timeTo: 'Closes',
-    open: 'Open',
-    priceName: 'Price name',
-    amount: 'Amount',
-    certificateName: 'Certificate',
-    accessibilityDescription: 'Accessibility',
-    accessibilityTypes: 'Accessibility types',
-    tags: 'Tags',
-    payload: 'Payload JSON',
-    actions: 'Actions',
-    createdAt: 'Created',
-    updatedAt: 'Updated',
-  },
-  actions: {
-    create: 'Create',
-    update: 'Save changes',
-    save: 'Save',
-    edit: 'Edit',
-    delete: 'Delete',
-    back: 'Back to list',
-    deleteConfirm: 'Delete this POI?',
-    clearMedia: 'Clear media',
-    add: 'Add',
-    remove: 'Remove',
-    searchAddress: 'Search address',
-    applySearchResult: 'Apply',
-    reverseGeocode: 'Resolve address from coordinates',
-    reverseGeocoding: 'Resolving address',
-    uploadMedia: 'Upload media',
-    uploadingMedia: 'Uploading media',
-  },
-  messages: {
-    loading: 'Loading POI.',
-    loadError: 'POI could not be loaded.',
-    missingContent: 'The POI could not be loaded.',
-    saveError: 'POI could not be saved.',
-    deleteError: 'POI could not be deleted.',
-    createSuccess: 'POI was created.',
-    updateSuccess: 'POI was updated.',
-    validationError: 'Please correct the highlighted fields.',
-    locationSearchError: 'Address search is currently unavailable. You can still edit the fields manually.',
-    locationSearchEmpty: 'No address matches were found.',
-    locationReverseGeocodeError: 'Reverse geocoding is currently unavailable.',
-    locationReverseGeocodeEmpty: 'No address was found for these coordinates.',
-    locationMapUnavailable: 'The map view is currently disabled for this instance.',
-    locationMapError: 'The map view could not be loaded. Manual editing remains available.',
-    mediaUploadError: 'The media upload failed.',
-    mediaUploadSuccess: 'The media was uploaded and can now be referenced.',
-  },
-  tabs: {
-    ariaLabel: 'Detail sections',
-    mobileLabel: 'Select section',
-  },
-  detailTabs: {
-    basis: { title: 'Basics', description: 'Master data and publication status.' },
-    location: { title: 'Location', description: 'Address, location data, and geo coordinates.' },
-    description: { title: 'Description', description: 'Editorial description fields.' },
-    contact: { title: 'Contact', description: 'Primary contact details.' },
-    openingHours: { title: 'Opening hours', description: 'Repeatable opening entries.' },
-    links: { title: 'Links', description: 'External references and related URLs.' },
-    operator: { title: 'Operator', description: 'Operating company and responsibility.' },
-    prices: { title: 'Prices', description: 'Price information and categories.' },
-    media: { title: 'Media & files', description: 'Teaser image and file references.' },
-    advanced: { title: 'Advanced data', description: 'Payload, certificates, and accessibility.' },
-    history: { title: 'History', description: 'Change and activity history.' },
-  },
-  cards: {
-    basis: {
-      identity: { title: 'Core data', description: 'Name, category, and activation state.' },
-      meta: { title: 'Metadata', description: 'Temporal information for the record.' },
-    },
-    content: {
-      descriptions: { title: 'Descriptions', description: 'Editorial descriptions for the POI.' },
-      contact: { title: 'Contact', description: 'Contact information for the POI.' },
-      location: { title: 'Location and address', description: 'Address data of the POI.' },
-      openingHours: { title: 'Opening hours', description: 'Current opening information.' },
-      links: { title: 'Web links', description: 'External references for the POI.' },
-      payload: { title: 'Additional data', description: 'Additional Mainserver data as JSON.' },
-    },
-    location: {
-      search: { title: 'Address search', description: 'Search an address and apply it to the POI.' },
-      address: { title: 'Location and address', description: 'Structured address data for the POI.' },
-      coordinates: { title: 'Coordinates', description: 'Geo coordinates of the POI.' },
-      map: { title: 'Map', description: 'Place or drag the marker to maintain coordinates.' },
-    },
-    description: {
-      text: { title: 'Descriptions', description: 'Editorial text fields for the POI.' },
-    },
-    contact: {
-      primary: { title: 'Contact', description: 'Contact information for the POI.' },
-    },
-    openingHours: {
-      entries: { title: 'Opening hours', description: 'Maintain multiple opening entries.' },
-    },
-    links: {
-      entries: { title: 'Web links', description: 'External references for the POI.' },
-    },
-    operator: {
-      details: { title: 'Operator', description: 'Responsible organization and reachability.' },
-    },
-    prices: {
-      entries: { title: 'Prices', description: 'Maintain multiple price entries.' },
-    },
-    media: {
-      references: { title: 'Media', description: 'Teaser image and file references.' },
-    },
-    advanced: {
-      payload: { title: 'Additional data', description: 'Payload, certificates, and accessibility.' },
-    },
-    settings: {
-      media: { title: 'Media', description: 'Teaser image for the overview.' },
-    },
-  },
-  history: {
-    empty: {
-      title: 'No history available yet.',
-      description: 'POI history entries will be connected in a later step.',
-    },
-  },
-  empty: { title: 'No POI yet', description: 'Create the first POI.' },
-  pagination: {
-    ariaLabel: 'POI pagination',
-    previous: 'Previous',
-    next: 'Next',
-    pageLabel: 'Page {{page}}',
-  },
-  values: {
-    notAvailable: 'Not available',
-    active: 'Yes',
-  },
-  validation: {
-    name: 'Name is required.',
-    webUrls: 'URLs must start with https://.',
-    categoryName: 'Category must be at most 128 characters.',
-    payload: 'Payload must be valid JSON.',
-  },
-});
+import { poiTranslationsDe } from './plugin.translations.de.js';
+import { poiTranslationsEn } from './plugin.translations.en.js';
 
 export const pluginPoiTranslations = {
   de: poiTranslationsDe,
