+++ packages/plugin-poi/src/poi.detail-media-tab.tsx
@@ -1,7 +1,8 @@
-import * as React from 'react';
-import { Alert, AlertDescription, Button, MediaReferenceField } from '@sva/studio-ui-react';
-import { Controller, useFieldArray, useFormContext } from 'react-hook-form';
+import { useFieldArray, useFormContext } from 'react-hook-form';
 
+import { PoiDetailMediaAttachments } from './poi.detail-media-attachments.js';
+import { PoiDetailMediaTeaserField } from './poi.detail-media-teaser-field.js';
+import { PoiDetailMediaUpload } from './poi.detail-media-upload.js';
 import { PoiDetailSectionCard } from './poi.detail-section-card.js';
 import type { PoiDetailFormValues } from './poi.detail-form.js';
 
@@ -21,94 +22,29 @@ export function PoiDetailMediaTab({
   pt: (key: string) => string;
 }>) {
   const { control } = useFormContext<PoiDetailFormValues>();
-  const { fields, append, remove } = useFieldArray({ control, name: 'media.attachments' });
-  const uploadInputId = React.useId();
-  const uploadInputRef = React.useRef<HTMLInputElement | null>(null);
-
-  const handleFileSelection = React.useCallback(
-    async (event: React.ChangeEvent<HTMLInputElement>) => {
-      const file = event.target.files?.[0];
-      event.target.value = '';
-      if (!file) {
-        return;
-      }
-      const assetId = await onUpload(file);
-      if (!assetId) {
-        return;
-      }
-      append({ assetId, label: '' });
-    },
-    [append, onUpload],
-  );
+  const { append, fields, remove } = useFieldArray({ control, name: 'media.attachments' });
 
   return (
     <PoiDetailSectionCard title={pt('cards.media.references.title')} description={pt('cards.media.references.description')}>
-      <div className="flex flex-wrap gap-3">
-        <input
-          ref={uploadInputRef}
-          id={uploadInputId}
-          type="file"
-          accept="image/*"
-          className="sr-only"
-          onChange={(event) => void handleFileSelection(event)}
-        />
-        <Button type="button" variant="outline" onClick={() => uploadInputRef.current?.click()} disabled={isUploading}>
-          {isUploading ? pt('actions.uploadingMedia') : pt('actions.uploadMedia')}
-        </Button>
-      </div>
-      {uploadError ? (
-        <Alert>
-          <AlertDescription>{uploadError}</AlertDescription>
-        </Alert>
-      ) : null}
-      {uploadSuccess ? (
-        <Alert>
-          <AlertDescription>{uploadSuccess}</AlertDescription>
-        </Alert>
-      ) : null}
-      <Controller
-        name="media.teaserImageAssetId"
-        control={control}
-        render={({ field }) => (
-          <MediaReferenceField
-            id="poi-teaser-image"
-            label={pt('fields.teaserImage')}
-            value={field.value || null}
-            options={mediaOptions}
-            onChange={(assetId) => field.onChange(assetId ?? '')}
-            placeholder={pt('fields.mediaPlaceholder')}
-            clearLabel={pt('actions.clearMedia')}
-          />
-        )}
+      <PoiDetailMediaUpload
+        isUploading={isUploading}
+        uploadError={uploadError}
+        uploadSuccess={uploadSuccess}
+        onUpload={onUpload}
+        onUploaded={(assetId) => append({ assetId, label: '' })}
+        pt={pt}
       />
-
-      {fields.map((field, index) => (
-        <div key={field.id} className="rounded-xl border border-border/60 p-4">
-          <Controller
-            name={`media.attachments.${index}.assetId`}
-            control={control}
-            render={({ field: mediaField }) => (
-              <MediaReferenceField
-                id={`poi-attachment-id-${index}`}
-                label={pt('fields.attachmentAsset')}
-                value={mediaField.value || null}
-                options={mediaOptions}
-                onChange={(assetId) => mediaField.onChange(assetId ?? '')}
-                placeholder={pt('fields.mediaPlaceholder')}
-                clearLabel={pt('actions.clearMedia')}
-              />
-            )}
-          />
-          <div className="mt-4 flex justify-end">
-            <Button type="button" variant="outline" onClick={() => remove(index)}>
-              {pt('actions.remove')}
-            </Button>
-          </div>
-        </div>
-      ))}
-      <Button type="button" variant="outline" onClick={() => append({ assetId: '', label: '' })}>
-        {pt('actions.add')}
-      </Button>
+      <PoiDetailMediaTeaserField
+        mediaOptions={mediaOptions}
+        pt={pt}
+      />
+      <PoiDetailMediaAttachments
+        fields={fields}
+        mediaOptions={mediaOptions}
+        onAppend={() => append({ assetId: '', label: '' })}
+        onRemove={remove}
+        pt={pt}
+      />
     </PoiDetailSectionCard>
   );
 }
