+++ packages/plugin-poi/src/poi.types.ts
@@ -1,79 +1,15 @@
-export type PoiAddress = {
-  readonly addition?: string;
-  readonly street?: string;
-  readonly zip?: string;
-  readonly city?: string;
-  readonly kind?: string;
-  readonly geoLocation?: {
-    readonly latitude?: number;
-    readonly longitude?: number;
-  };
-};
-
-export type PoiContact = {
-  readonly firstName?: string;
-  readonly lastName?: string;
-  readonly phone?: string;
-  readonly fax?: string;
-  readonly email?: string;
-  readonly webUrls?: readonly PoiWebUrl[];
-};
-
-export type PoiWebUrl = {
-  readonly url: string;
-  readonly description?: string;
-};
-
-export type PoiLocation = {
-  readonly name?: string;
-  readonly department?: string;
-  readonly district?: string;
-  readonly regionName?: string;
-  readonly state?: string;
-  readonly geoLocation?: {
-    readonly latitude?: number;
-    readonly longitude?: number;
-  };
-};
-
-export type PoiOperatingCompany = {
-  readonly name?: string;
-  readonly address?: PoiAddress;
-  readonly contact?: PoiContact;
-};
-
-export type PoiPriceInformation = {
-  readonly name?: string;
-  readonly amount?: number;
-  readonly groupPrice?: boolean;
-  readonly ageFrom?: number;
-  readonly ageTo?: number;
-  readonly minAdultCount?: number;
-  readonly maxAdultCount?: number;
-  readonly minChildrenCount?: number;
-  readonly maxChildrenCount?: number;
-  readonly description?: string;
-  readonly category?: string;
-};
-
-export type PoiMediaContent = {
-  readonly captionText?: string;
-  readonly copyright?: string;
-  readonly height?: number;
-  readonly width?: number;
-  readonly contentType?: string;
-  readonly sourceUrl?: PoiWebUrl;
-};
-
-export type PoiCertificate = {
-  readonly name: string;
-};
-
-export type PoiAccessibilityInformation = {
-  readonly description?: string;
-  readonly types?: string;
-  readonly urls?: readonly PoiWebUrl[];
-};
+import type {
+  PoiAccessibilityInformation,
+  PoiAddress,
+  PoiCertificate,
+  PoiContact,
+  PoiLocation,
+  PoiMediaContent,
+  PoiOpeningHour,
+  PoiOperatingCompany,
+  PoiPriceInformation,
+  PoiWebUrl,
+} from './poi.content.types.js';
 
 export type PoiListQuery = {
   readonly page: number;
@@ -90,18 +26,6 @@ export type PoiPagination = {
 export type PoiListResult = {
   readonly data: readonly PoiContentItem[];
   readonly pagination: PoiPagination;
-};
-
-export type PoiOpeningHour = {
-  readonly weekday?: string;
-  readonly dateFrom?: string;
-  readonly dateTo?: string;
-  readonly timeFrom?: string;
-  readonly timeTo?: string;
-  readonly sortNumber?: number;
-  readonly open?: boolean;
-  readonly useYear?: boolean;
-  readonly description?: string;
 };
 
 export type PoiFormInput = {
